NF: move Leech.runHook to AbstractSched
