+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/SchedV2.java
@@ -39,7 +39,6 @@ import com.ichi2.libanki.Decks;
 import com.ichi2.libanki.Note;
 import com.ichi2.libanki.Utils;
 
-import com.ichi2.libanki.hooks.Leech;
 import com.ichi2.libanki.utils.SystemTime;
 import com.ichi2.libanki.utils.Time;
 import com.ichi2.utils.JSONArray;
@@ -1726,8 +1725,8 @@ public class SchedV2 extends AbstractSched {
             }
             // notify UI
             if (mContextReference != null) {
-                Context context = mContextReference.get();
-                new Leech().runHook(card, context);
+                Activity context = mContextReference.get();
+                leech(card, context);
             }
             return true;
         }
